Mike commit

+++ Evershop.Tests/Evershop.Tests.API/ApiTests.cs
@@ -11,6 +11,8 @@ namespace Evershop.Tests.API
         private RestClient _client;
         private string _sid;
         private CookieCollection _cookies;
+        private string uuid;
+        private App app;
 
         [SetUp]
         public void Setup()
@@ -40,18 +42,18 @@ namespace Evershop.Tests.API
                 request.AddCookie(_cookies[i].Name, _cookies[i].Value, _cookies[i].Path, _cookies[i].Domain);
             }
             request.AddHeader("Authorization", $"Bearer {_sid}");
-            
+
             var product = new Product()
             {
                 Name = "Create product test 1",
                 ProductId = "",
-                Sku = "CPT-1",
+                Sku = "CPT-2",
                 Price = "12.25",
                 Weight = "0.28",
                 CategoryId = "3",
                 TaxClass = "",
                 Description = "<p>Create product test 1 description</p>",
-                UrlKey = "product1",
+                UrlKey = "product2",
                 MetaTitle = "",
                 MetaKeywords = "",
                 MetaDescription = "",
@@ -93,9 +95,9 @@ namespace Evershop.Tests.API
         [Test]
         public async Task CreateProduct2()
         {
-            var app = new App();
+            app = new App();
 
-            
+
             // Arrange
             var request = new RestRequest("http://localhost:3000/api/products", Method.Post);
 
@@ -137,27 +139,35 @@ namespace Evershop.Tests.API
             //var response = _client.Execute(request);
             var response = await app.ApiClient.PostAsync(request);
             // Assert
-            
-            var jsonResponse = JObject.Parse(response.Content);
-            Assert.That(response.StatusCode, Is.EqualTo(HttpStatusCode.OK), jsonResponse.ToString());
+
+            var jsonResponse = JObject.Parse(response.Response.Content);
+            uuid = jsonResponse["data"]["uuid"].ToString();
+            Assert.That(response.Response.StatusCode, Is.EqualTo(HttpStatusCode.OK), jsonResponse.ToString());
             Assert.IsNotNull(jsonResponse["data"]);  // Check if new ID was returned
             Assert.IsNotNull(jsonResponse["data"]["product_description_id"]);  // Check if new ID was returned
             await Console.Out.WriteLineAsync(jsonResponse["data"]["product_description_id"].ToString());
-
-            // Delete
-            string uuid = jsonResponse["data"]["uuid"].ToString();
-            request = new RestRequest($"http://localhost:3000/api/products/{uuid}", Method.Delete);
-            for (int i = 0; i < _cookies.Count; i++)
-            {
-                request.AddCookie(_cookies[i].Name, _cookies[i].Value, _cookies[i].Path, _cookies[i].Domain);
-            }
-            request.AddHeader("Authorization", $"Bearer {_sid}");
-            response = await app.ApiClient.DeleteAsync(request);
-
-            Assert.That(response.StatusCode, Is.EqualTo(HttpStatusCode.OK));
+            Assert.That(response.ExecutionTime, Is.LessThan(TimeSpan.FromMilliseconds(150)));
         }
 
-    }
+        [TearDown]
+        public async Task TearDownAsync()
+        {
+            if(uuid != null)
+            {
+                // Delete
+                var request = new RestRequest($"http://localhost:3000/api/products/{uuid}", Method.Delete);
+                for (int i = 0; i < _cookies.Count; i++)
+                {
+                    request.AddCookie(_cookies[i].Name, _cookies[i].Value, _cookies[i].Path, _cookies[i].Domain);
+                }
+                request.AddHeader("Authorization", $"Bearer {_sid}");
+                var response = await app.ApiClient.DeleteAsync(request);
 
+                Assert.That(response.Response.StatusCode, Is.EqualTo(HttpStatusCode.OK));
+
+            }
+
+        }
+    }
 
 }
